+++ CHANGELOG.md
@@ -4,13 +4,15 @@ This is the changelog for Cldr_Currencies v2.5.0 released on ________.  For olde
 
 ### Enhancements
 
-* Add `Cldr.Currency.currency_from_locale/1`
+* Add `Cldr.Currency.currency_from_locale/{1, 2}`
 
-* Add `Cldr.Currency.current_currency_for_locale/1`
+* Add `Cldr.Currency.current_currency_from_locale/1`
 
-* Add `Cldr.Currency.current_currency_for_territory/1`
+* Add `Cldr.Currency.current_currency_from_territory/1`
 
 * Add `Cldr.Currency.currency_format_from_locale/1`
+
+* Add `<backend>.currency_from_locale/{1, 2}`
 
 # Changelog for Cldr_Currencies v2.4,1